==> Scripts/2. Importing libraries and Python data types.ipynb <==
# 01. Importing libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Data types

x = 2

x

print(x)

type(x)

y=5.678

type(y)

y=5.678

type(y)

z='Brazil'

type(z)

type('Mania')

type(true)

type(True)

print(x)

x+4

x+y

x=x+4

print(x)

a='Hello,'

b='Marley'

a+b

n='potato'

n*10

x==y

type(x==y)

# Task - 4.2 Jupyter Fundamentals & Python Data Types

## Question 7

x=38

y=62

x+y

(y*2)-24

x=y+45

(x+y)-69

## Question 8

x=37.934

y=21.35

x/y

## Question 9

x='United '

y='States'

x+y

## Question 10

a='What goes around, '

b='comes around.'

a+b

c='Python is a '

d='programming language.'

c+d

## Bonus

a='Python'

a*100
