==> Scripts/3. Data import and descriptive analysis.ipynb <==
## Task - 4.3 Data Import and Descriptive Analysis

# Import libraries
import pandas as pd
import numpy as np
import os

df = pd.read_csv(r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024/02. Data/Original Data/orders.csv', index_col = False)

df=pd.read_csv(r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024/02. Data/Original Data/orders.csv', index_col=False)

# Import libraries
import pandas as pd
import numpy as np
import os

df=pd.read_csv(r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024/02. Data/Original Data/orders.csv', index_col=False)

## Question 4

path=r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

path

df=pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), index_col=False)

df.head()

df.head(30)

df.tail()

df.shape

df.info()

df.dtypes

## Question 5

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), usecols = vars_list)

## Question 6

df.head()

## Question 7

df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

## Question 8

df_prods.head(20)

df_prods.tail(35)

## Question 9

df_prods.columns

## Question 10

df_prods.shape

## Question 11(1)

df_prods.describe()

### Answer: max value of aisle_id is 134

## Question 11(2)

df_prods.dtypes

### Answer: data type of "department_id" is int64
