==> Scripts/4. Data Wrangling & Subsetting.ipynb <==
## 4.4 Data Wrangling & Subsetting

#import libraries
import pandas as pd
import numpy as np
import os

path=r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

df_ords = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

# Dropping eval_set column from orders.csv
df_ords.drop(columns = ['eval_set'])

# Checking for missing values 
df_ords['days_since_prior_order'].value_counts (dropna = False)

# Renaming 'order_dow' column
df_ords.rename (columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

df_ords.describe()

# Changing 'order_id's data type
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtypes

df_dep = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing data
df_dep.T

df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

# Take the first rew of df_dep_t for the header
new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Set the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

df_prods

# Creating a data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()

## Task - 4.4 

### Question 2

df_ords.info()

# Changing 'user_id's data type
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

### Question 3

df_ords

# Renaming 'order_number' column
df_ords.rename (columns = {'order_number' : 'number_of_orders'})

### Question 4

df_ords['order_hour_of_day'].value_counts (dropna = False)

# Busiest hour for placing orders is 10 AM in the mornings.

### Question 5

df_dep_t_new

# The value of 4 in the "department_id" column within the df_prods is "Produce"

### Question 6

df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast

### Question 7

df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_party

### Question 8

# 7650 rows 

### Question 9

# Info about the user_id "1"
df_user = df_ords[df_ords['user_id']=='1']

df_user

### Question 10

# Deatils about user_id 1's order behaviour 
df_user.describe()

# This user has made a total of 11 orders and the user usually place orders on Wednesdays at 7 or 8 AM in the morning. 
# The user's ordering frequency varies but the average time between the user's orders is 19 days.

### Question 12

# Exporting df_ords dataframe
df_ords.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_wrangled.csv'))

### Question 13

# Exporting df_dep_t_new dataframe
df_dep_t_new.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'departments_wrangled.csv'))
