==> Scripts/5. Data Consistency Checks.ipynb <==
# Importing libraries
import pandas as pd
import numpy as np
import os

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

df_ords = pd.read_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords.head()

df_ords = df_ords.drop(columns=['Unnamed: 0'])

df_ords.head()

df_ords.describe()

# Create a dataframe
df_test = pd.DataFrame()

df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Check for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

df_prods.isnull()

# Cheking for missing values
df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

# Removing missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Checking for duplicate values
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

# Removing duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

## Task 4.4 Data Consistency Checks

### Question 2

df_ords.describe()

# order_number: The min value is 1 and the max value is 100 which sounds a bit high for one user so it's worth investigating further.
    
# orders_day_of_week: The min value is 0 and the max value is 6, which is expected for days of the week.

# order_hour_of_day: The min value is 0 and the max value is 23, which is expected for hours of the day.

# days_since_prior_order: The min value is 0 and the max value is 30. It looks reasonable but it's worth investigating further why the the gap between orders are so long.

### Question 3

df_ords.info()

# Looking for mix types
for column in df_ords.columns:
 print(pd.api.types.infer_dtype(df_ords[column]))

# I found this code on the internet because I wanted to see the if eval_set is a mix type but the result is string

# This is the code that was used in the reading material but I got nothing as a result and I don't know how to interpret the error message.
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

### Question 5

# Cheking for missing values
df_ords.isnull().sum()

# There are 206209 missing values in the "days_since_prior_order" column.

# One reason for the missing values is because of first time customers who are placing their first order 
# wouldn't have a "days_since_prior_order" value.

# Another reason could be one time or infrequent orders if a customer places an order 
# and then doesn't place any order for a considerable amount of time.

### Question 6

# One way to tackle the missing values is to create a new variable that acts like a flag based on the missing value:
# in this case we can create a new column for new customers that has no prior order date 
# or create a column to seperate more frequent users than irregular users.

# Another way to tackle the missing values is to impute them and this would be the method I would choose. I would change NaN values in the
# days_since_prior_order to 0 as it was 0 days between their prior order.

# Last option would be to remove them however it wouldn't be the method I would choose because the missing values. 
# seem to have important info about customer behaviour.

df_ords['days_since_prior_order'].fillna(0, inplace = True)

df_ords.isnull().sum()

### Question 7

# Checking for duplicate values 
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

# There are no duplicate values found 

df_ords.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_cleaned.csv'))

df_prods.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'products_cleaned.csv'))
