==> Scripts/6. Combining & Exporting Data (Task).ipynb <==
import pandas as pd
import numpy as np
import os

### Question 3

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

# Pickle file imported
df_merged = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_merged.head()

### Question 4

# Shape of merged data
df_merged.shape

### Question 5

df_prods_cleaned = pd.read_csv(os.path.join(path, '02. Data', 'Prepared Data', 'products_cleaned.csv'))

df_prods_cleaned.head()

df_prods = df_prods_cleaned.drop(['Unnamed: 0'], axis = 1)

df_prods.head()

#  Combining the orders_products_combined dataframe with products data set.
df_ords_prods_merge = df_merged.merge(df_prods, on = 'product_id')

### Question 6

df_ords_prods_merge.head()

# Merge flag frequency
df_ords_prods_merge['_merge'].value_counts()

### Question 7

# Exporting the merged data to pickle format
df_ords_prods_merge.to_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_merged.pkl'))
