==> Scripts/7. Deriving New Variables.ipynb <==
import pandas as pd
import numpy as np
import os

# Define function

def add_num(x,y):
    sum = x + y
    return sum 

# Run function

add_num(9,11)

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Create a subset

df = df_ords_prods_merged[:1000000]

df.shape

# Defining price range function

def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else: return 'Not enough data'

# Apply function

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'
df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

for x in range(30, 45):
    print('My age is %d' % (x))

df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

result = []
for value in df_ords_prods_merged['orders_day_of_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least Busy')
    else:
        result.append('Regularly busy')

result

df_ords_prods_merged['busiest day'] = result

df_ords_prods_merged['busiest day'].value_counts(dropna = False)

## Task - 4.7 Deriving New Variables

### Question 2

# Creating new labels 

result = []
for value in df_ords_prods_merged['orders_day_of_week']:
    if value == 0 or value == 1:  
        result.append('Busiest days')
    elif value == 3 or value == 4: 
        result.append('Slowest days')
    else:
        result.append('Regularly busy')

### Question 3

df_ords_prods_merged['busiest days'] = result

df_ords_prods_merged['busiest days'].value_counts(dropna = False)

### Question 4

# Finding busiest hour of the day

df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

result = []
for value in df_ords_prods_merged['order_hour_of_day']:
    if value in [10, 11, 14, 15, 13, 12, 16, 9]:  
        result.append('Most orders')
    elif value in [3, 4, 2, 5, 1, 0, 6, 23]: 
        result.append('Fewest orders')
    else:
        result.append('Average orders')

### Question 5

df_ords_prods_merged['busiest_period_of_day'] = result

df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

#### The following part, I found online because I wasn't sure about how to interpret "busiest hours".

percentiles = np.percentile(df_ords_prods_merged['order_hour_of_day'], [25, 75])

most_orders_hours = df_ords_prods_merged['order_hour_of_day'] >= percentiles[1]
fewest_orders_hours = df_ords_prods_merged['order_hour_of_day'] <= percentiles[0]
average_orders_hours = ~most_orders_hours & ~fewest_orders_hours

df_ords_prods_merged['hour_order_volume'] = np.where(most_orders_hours, 'Most orders',
                                                     np.where(fewest_orders_hours, 'Fewest orders', 'Average orders'))

df_ords_prods_merged['hour_order_volume'].value_counts(dropna = False)

df_ords_prods_merged.to_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_merged_new.pkl'))
