==> Scripts/8. Grouping Data & Aggregating Variables.ipynb <==
import pandas as pd
import numpy as np
import os

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_merged_new.pkl'))

ords_prods_merge.head()

# Creating a subset called df of the first one million entries

df = ords_prods_merge[:1000000]

# Check the shape of the new df

df.shape

df.head()

# Splitting the data into groups

df.groupby('product_name')

# Split the data into groups based on “department_id” and aggregating to mean for “order_number” 

df.groupby('department_id').agg({'order_number': ['mean']})

df.groupby('department_id')['order_number'].mean()

# Performing Multiple Aggregations

df.groupby('department_id').agg({'order_number': ['mean', 'max', 'min']})

# Aggregating Data with transform()

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.head(20)

# Deriving Columns with loc()

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

## Task - 4.8 Grouping Data and Aggregating Variables

### Question 2

### Question 3

### Question 5

ords_prods_merge.groupby('loyalty_flag')['prices'].mean()

ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'max', 'min']})

# It looks like loyal customers usually to spend less on average compared to new or regular customers

### Question 6

# Split data into groups based on “user_id” and aggregating to mean for “prices” 

ords_prods_merge.groupby('user_id').agg({'prices': ['mean']})

# Aggregating Data with transform()

ords_prods_merge['avg_spending'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.head()

#  Creating a spending flag

ords_prods_merge.loc[ords_prods_merge['avg_spending'] < 10, 'spending_flag'] = 'Low spender'
ords_prods_merge.loc[ords_prods_merge['avg_spending'] >= 10, 'spending_flag'] = 'High spender'

ords_prods_merge['spending_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id', 'avg_spending', 'spending_flag']].head (100)

### Question 7

# Split data into groups based on “user_id” and aggregating to median for “days_since_prior_order” 

ords_prods_merge.groupby('user_id').agg({'days_since_prior_order': ['median']})

# Aggregating Data with transform()

ords_prods_merge['order_frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.head()

#  Creating an order frequency flag

ords_prods_merge.loc[ords_prods_merge['order_frequency'] > 20, 'order_frequency_flag'] = 'Non-frequent customer'
ords_prods_merge.loc[(ords_prods_merge['order_frequency'] <= 20) & (ords_prods_merge['order_frequency'] > 10), 'order_frequency_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['order_frequency'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

ords_prods_merge[['user_id', 'order_frequency', 'order_frequency_flag']].head (50)

ords_prods_merge.to_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'ords_prods_merge_aggregate.pkl'))
