==> Scripts/9. Data Visualisation with Python (Part 1).ipynb <==
### Question 3

# Import libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

pd.set_option('display.max_columns', None)

# Import customer dataset

path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

customers = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'customers.csv'), index_col = False)

customers

# Checking for shape

customers.shape

# Cheking for missing values

customers.isnull().sum()

# Cheking out the missing values in first name column

customers[customers['First Name'].isnull()]

# Exploring the data types

customers.info()

# Cheking data contents with describe

customers.describe()

# Cheking Gender distribution

customers['Gender'].value_counts()

# Checking state distribution

customers['STATE'].value_counts(dropna = False)

# Checking age distibution

customers['Age'].value_counts(dropna = False)

# Cheking date_joined

customers['date_joined'].value_counts(dropna = False)

# changing Surnam column

customers_2 = customers.rename(columns={'Surnam' : 'Last Name'})

customers_2.head()

customers_3 = customers_2.rename(columns={'STATE' : 'State'})

customers_3.head()

# Cheking for duplicates

dups = customers_3[customers_3.duplicated()]

dups.head()

### Question 6

# Import merged dataset 
path = r'/Users/selcenayranci/Desktop/4th Achievement -  Instacart Basket Analysis March 2024'

ords_prods_merged = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'ords_prods_merge_aggregate.pkl'))

ords_prods_merged.info()

#  Combining the ords_prods_merged dataframe with customers data set.

ords_prods_cus = ords_prods_merged.merge(customers_3, on = 'user_id')

ords_prods_cus.head()

ords_prods_cus.columns

# Removing unnecessary columns

ords_prods_cus_subset = ords_prods_cus[['order_id', 'user_id', 'order_number', 'orders_day_of_week',
       'order_hour_of_day', 'days_since_prior_order', 'product_id', 'product_name',
       'department_id', 'prices', 'price_range_loc', 'busiest day',
       'busiest_period_of_day', 'hour_order_volume', 'busiest days',
       'max_order', 'loyalty_flag', 'avg_spending', 'spending_flag',
       'order_frequency', 'order_frequency_flag', 'Gender', 'State', 'Age', 'n_dependants', 'fam_status',
       'income']]

ords_prods_cus_subset.head()

# Exporting the dataframe

ords_prods_cus_subset.to_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'ords_prods_cus.pkl'))
